# Conflicts:
#	libraries/AP_Button/AP_Button.cpp

+++ Tools/autotest/pysim/util.py
@@ -24,6 +24,10 @@ else:
 
 RADIUS_OF_EARTH = 6378100.0  # in meters
 
+
+# List of open terminal windows for macosx
+windowID = []
+
 def m2ft(x):
     """Meters to feet."""
     return float(x) / 0.3048
@@ -92,10 +96,12 @@ def relwaf():
     return "./modules/waf/waf-light"
 
 
-def waf_configure(board, j=None, debug=False, extra_args=[]):
+def waf_configure(board, j=None, debug=False, math_check_indexes=False, extra_args=[]):
     cmd_configure = [relwaf(), "configure", "--board", board]
     if debug:
         cmd_configure.append('--debug')
+    if math_check_indexes:
+        cmd_configure.append('--enable-math-check-indexes')
     if j is not None:
         cmd_configure.extend(['-j', str(j)])
     pieces = [shlex.split(x) for x in extra_args]
@@ -108,12 +114,16 @@ def waf_clean():
     run_cmd([relwaf(), "clean"], directory=topdir(), checkfail=True)
 
 
-def build_SITL(build_target, j=None, debug=False, board='sitl', clean=True, configure=True, extra_configure_args=[]):
+def build_SITL(build_target, j=None, debug=False, board='sitl', clean=True, configure=True, math_check_indexes=False, extra_configure_args=[]):
     """Build desktop SITL."""
 
     # first configure
     if configure:
-        waf_configure(board, j=j, debug=debug, extra_args=extra_configure_args)
+        waf_configure(board,
+                      j=j,
+                      debug=debug,
+                      math_check_indexes=math_check_indexes,
+                      extra_args=extra_configure_args)
 
     # then clean
     if clean:
@@ -167,6 +177,9 @@ def pexpect_close(p):
     global close_list
 
     ex = None
+    if p is None:
+        print("Nothing to close")
+        return
     try:
         p.kill(signal.SIGTERM)
     except IOError as e:
@@ -227,6 +240,13 @@ def kill_screen_gdb():
     cmd = ["screen", "-X", "-S", "ardupilot-gdb", "quit"]
     subprocess.Popen(cmd)
 
+def kill_mac_terminal():
+    global windowID
+    for window in windowID:
+        cmd = ("osascript -e \'tell application \"Terminal\" to close "
+            "(window(get index of window id %s))\'" % window)
+        os.system(cmd)
+
 def start_SITL(binary,
                valgrind=False,
                gdb=False,
@@ -235,7 +255,7 @@ def start_SITL(binary,
                home=None,
                model=None,
                speedup=1,
-               defaults_file=None,
+               defaults_filepath=None,
                unhide_parameters=False,
                gdbserver=False,
                breakpoints=[],
@@ -243,6 +263,10 @@ def start_SITL(binary,
                vicon=False,
                customisations=[],
                lldb=False):
+
+    if model is None:
+        raise ValueError("model must not be None")
+
     """Launch a SITL instance."""
     cmd = []
     if valgrind and os.path.exists('/usr/bin/valgrind'):
@@ -281,7 +305,9 @@ def start_SITL(binary,
             f.write("disable\n")
         f.write("r\n")
         f.close()
-        if os.environ.get('DISPLAY'):
+        if sys.platform == "darwin" and os.getenv('DISPLAY'):
+            cmd.extend(['gdb', '-x', '/tmp/x.gdb', '--args'])
+        elif os.environ.get('DISPLAY'):
             cmd.extend(['xterm', '-e', 'gdb', '-x', '/tmp/x.gdb', '--args'])
         else:
             cmd.extend(['screen',
@@ -299,7 +325,9 @@ def start_SITL(binary,
         f.write("settings set target.process.stop-on-exec false\n")
         f.write("process launch\n")
         f.close()
-        if os.environ.get('DISPLAY'):
+        if sys.platform == "darwin" and os.getenv('DISPLAY'):
+            cmd.extend(['lldb', '-s', '/tmp/x.lldb', '--'])
+        elif os.environ.get('DISPLAY'):
             cmd.extend(['xterm', '-e', 'lldb', '-s','/tmp/x.lldb', '--'])
         else:
             raise RuntimeError("DISPLAY was not set")
@@ -311,19 +339,49 @@ def start_SITL(binary,
         cmd.append('-S')
     if home is not None:
         cmd.extend(['--home', home])
-    if model is not None:
-        cmd.extend(['--model', model])
+    cmd.extend(['--model', model])
     if speedup != 1:
         cmd.extend(['--speedup', str(speedup)])
-    if defaults_file is not None:
-        cmd.extend(['--defaults', defaults_file])
+    if defaults_filepath is not None:
+        if type(defaults_filepath) == list:
+            defaults_filepath = ",".join(defaults_filepath)
+        cmd.extend(['--defaults', defaults_filepath])
     if unhide_parameters:
         cmd.extend(['--unhide-groups'])
     if vicon:
         cmd.extend(["--uartF=sim:vicon:"])
     cmd.extend(customisations)
 
-    if gdb and not os.getenv('DISPLAY'):
+    if (gdb or lldb) and sys.platform == "darwin" and os.getenv('DISPLAY'):
+        global windowID
+        # on MacOS record the window IDs so we can close them later
+        atexit.register(kill_mac_terminal)
+        child = None
+        mydir = os.path.dirname(os.path.realpath(__file__))
+        autotest_dir = os.path.realpath(os.path.join(mydir, '..'))
+        runme = [os.path.join(autotest_dir, "run_in_terminal_window.sh"), 'mactest']
+        runme.extend(cmd) 
+        print(runme)
+        print(cmd)
+        out = subprocess.Popen(runme, stdout=subprocess.PIPE).communicate()[0]
+        out = out.decode('utf-8')
+        p = re.compile('tab 1 of window id (.*)')
+
+        tstart = time.time()
+        while time.time() - tstart < 5:
+            tabs = p.findall(out)
+
+            if len(tabs) > 0:
+                break
+
+            time.sleep(0.1)
+        # sleep for extra 2 seconds for application to start
+        time.sleep(2)
+        if len(tabs) > 0:
+            windowID.append(tabs[0])
+        else:
+            print("Cannot find %s process terminal" % binary)
+    elif gdb and not os.getenv('DISPLAY'):
         subprocess.Popen(cmd)
         atexit.register(kill_screen_gdb)
         # we are expected to return a pexpect wrapped around the
@@ -334,13 +392,14 @@ def start_SITL(binary,
                              logfile=sys.stdout,
                              encoding=ENCODING,
                              timeout=5)
+    else:
+        print("Running: %s" % cmd_as_shell(cmd))
 
-    print("Running: %s" % cmd_as_shell(cmd))
 
-    first = cmd[0]
-    rest = cmd[1:]
-    child = pexpect.spawn(first, rest, logfile=sys.stdout, encoding=ENCODING, timeout=5)
-    pexpect_autoclose(child)
+        first = cmd[0]
+        rest = cmd[1:]
+        child = pexpect.spawn(first, rest, logfile=sys.stdout, encoding=ENCODING, timeout=5)
+        pexpect_autoclose(child)
     # give time for parameters to properly setup
     time.sleep(3)
     if gdb or lldb:
